requirements.txt
pytorch
einops
numpy
torchvision
